refactor(fsd): Step 2 — shared layer migration
Move shared primitives from hybrid src/ layout into FSD shared/ layer:
- components/ui/* (12 files) → shared/ui/*/index.tsx
- components/common/* (6 files) → shared/ui/{table,empty-state,error-state,skeleton,pagination}/
- api/{client,queryClient}.ts → shared/api/
- lib/utils.ts (cn only) → shared/lib/cn.ts
- styles/index.css → shared/styles/globals.css
- hooks/useDrawer.ts → shared/hooks/useDisclosure.ts (renamed)
- mocks/** → test/mocks/**
- shared/ui/AttachmentZone/ → shared/ui/attachment-zone/ (kebab-case)
- Add shared/config/env.ts

Update all consumer imports to @shared/* aliases (26 files).
typecheck clean, 44 test files, 374 tests pass.
