+++ frontend/src/components/layout/Sidebar.tsx
@@ -9,7 +9,7 @@ import {
   Settings,
   Code2,
 } from 'lucide-react';
-import { cn } from '../../lib/utils';
+import { cn } from '@shared/lib/cn';
 import { useRole } from '../../hooks/useRole';
 
 interface NavItem {
