+++ frontend/src/components/voc/__tests__/VocListHeader.test.tsx
@@ -82,7 +82,10 @@ describe('VocListHeader', () => {
   // The container shadow lives outside the marker block (Phase B precedent allows
   // light-dark wrapped color-space literals in CSS for prototype-fidelity shadows).
   it('C-6 CSS block (between START/END markers) contains no hex or raw OKLCH', () => {
-    const css = readFileSync(resolve(__dirname, '..', '..', '..', 'styles', 'index.css'), 'utf-8');
+    const css = readFileSync(
+      resolve(__dirname, '..', '..', '..', 'shared', 'styles', 'globals.css'),
+      'utf-8',
+    );
     const start = css.indexOf('=== C-6 VocListHeader styles START ===');
     const end = css.indexOf('=== C-6 VocListHeader styles END ===');
     expect(start).toBeGreaterThan(-1);
