+++ frontend/src/components/voc/__tests__/VocRow.test.tsx
@@ -120,7 +120,10 @@ describe('VocRow', () => {
   });
 
   it('C-7 CSS block (between START/END markers) has no hex or raw oklch', () => {
-    const css = readFileSync(resolve(__dirname, '..', '..', '..', 'styles', 'index.css'), 'utf-8');
+    const css = readFileSync(
+      resolve(__dirname, '..', '..', '..', 'shared', 'styles', 'globals.css'),
+      'utf-8',
+    );
     const start = css.indexOf('=== C-7 VocRow styles START ===');
     const end = css.indexOf('=== C-7 VocRow styles END ===');
     expect(start).toBeGreaterThan(-1);
