+++ frontend/src/features/voc/components/VocCommentList.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
-import { Button } from '../../../components/ui/button';
-import { Textarea } from '../../../components/ui/textarea';
+import { Button } from '@shared/ui/button';
+import { Textarea } from '@shared/ui/textarea';
 
 export interface Comment {
   id: string;
