+++ frontend/src/features/voc/components/VocInternalNotes.tsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
-import { Button } from '../../../components/ui/button';
-import { Textarea } from '../../../components/ui/textarea';
-import { LoadingState } from '../../../components/common/LoadingState';
+import { Button } from '@shared/ui/button';
+import { Textarea } from '@shared/ui/textarea';
+import { LoadingState } from '@shared/ui/skeleton';
 import type { InternalNote } from '../../../../../shared/contracts/voc';
 import type { Role } from '../../../../../shared/contracts/common';
