+++ frontend/src/features/voc/components/VocStatusPriorityGrid.tsx
@@ -1,10 +1,4 @@
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '../../../components/ui/select';
+import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '@shared/ui/select';
 import { VocStatus, VocPriority, type VocUpdate } from '../../../../../shared/contracts/voc';
 import type { Voc } from '../../../../../shared/contracts/voc/entity';
