+++ frontend/src/shared/hooks/useDisclosure.ts
@@ -1,8 +1,8 @@
 import { useContext } from 'react';
-import { VOCDrawerContext, VOCDrawerContextValue } from '../contexts/VOCDrawerContext';
+import { VOCDrawerContext, VOCDrawerContextValue } from '../../contexts/VOCDrawerContext';
 
-export function useDrawer(): VOCDrawerContextValue {
+export function useDisclosure(): VOCDrawerContextValue {
   const ctx = useContext(VOCDrawerContext);
-  if (!ctx) throw new Error('useDrawer must be used within VOCDrawerProvider');
+  if (!ctx) throw new Error('useDisclosure must be used within VOCDrawerProvider');
   return ctx;
 }
