+++ frontend/src/shared/ui/alert/index.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { cn } from '../../lib/utils';
+import { cn } from '@shared/lib/cn';
 
 type AlertVariant = 'default' | 'destructive' | 'warning' | 'success';
