+++ frontend/src/shared/ui/empty-state/EmptyState.tsx
@@ -1,6 +1,6 @@
 import type { LucideIcon } from 'lucide-react';
-import { cn } from '../../lib/utils';
-import { Button } from '../ui/button';
+import { cn } from '@shared/lib/cn';
+import { Button } from '@shared/ui/button';
 
 interface EmptyStateProps {
   icon?: LucideIcon;
