+++ frontend/src/shared/ui/input/index.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { cn } from '../../lib/utils';
+import { cn } from '@shared/lib/cn';
 
 export type InputProps = React.InputHTMLAttributes<HTMLInputElement>;
