+++ frontend/src/shared/ui/popover/index.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import * as PopoverPrimitive from '@radix-ui/react-popover';
-import { cn } from '../../lib/utils';
+import { cn } from '@shared/lib/cn';
 
 const Popover = PopoverPrimitive.Root;
 const PopoverTrigger = PopoverPrimitive.Trigger;
