+++ frontend/src/shared/ui/select/index.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import * as SelectPrimitive from '@radix-ui/react-select';
 import { Check, ChevronDown, ChevronUp } from 'lucide-react';
-import { cn } from '../../lib/utils';
+import { cn } from '@shared/lib/cn';
 
 const Select = SelectPrimitive.Root;
 const SelectGroup = SelectPrimitive.Group;
