+++ frontend/src/shared/ui/skeleton/LoadingState.tsx
@@ -1,4 +1,4 @@
-import { cn } from '../../lib/utils';
+import { cn } from '@shared/lib/cn';
 
 interface LoadingStateProps {
   label?: string;
