+++ frontend/src/shared/ui/tabs/index.tsx
@@ -1,6 +1,6 @@
 import * as React from 'react';
 import * as TabsPrimitive from '@radix-ui/react-tabs';
-import { cn } from '../../lib/utils';
+import { cn } from '@shared/lib/cn';
 
 const Tabs = TabsPrimitive.Root;
