+++ frontend/src/shared/ui/textarea/index.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import { cn } from '../../lib/utils';
+import { cn } from '@shared/lib/cn';
 
 export type TextareaProps = React.TextareaHTMLAttributes<HTMLTextAreaElement>;
